Use bfloat16 as default dtype for qwen_image_edit_lora_collection
The QwenImageEditPlusPipeline model weights are natively in bfloat16.
Using float32 caused a dtype mismatch error during inference.

+++ qwen_image_edit_lora_collection/pytorch/loader.py
@@ -113,7 +113,7 @@ class ModelLoader(ForgeModel):
         Returns:
             QwenImageTransformer2DModel with LoRA weights fused.
         """
-        dtype = dtype_override if dtype_override is not None else torch.float32
+        dtype = dtype_override if dtype_override is not None else torch.bfloat16
 
         pipe = QwenImageEditPlusPipeline.from_pretrained(
             self._variant_config.pretrained_model_name,
@@ -136,7 +136,7 @@ class ModelLoader(ForgeModel):
 
         Returns a dict matching QwenImageTransformer2DModel.forward() signature.
         """
-        dtype = kwargs.get("dtype_override", torch.float32)
+        dtype = kwargs.get("dtype_override", torch.bfloat16)
         batch_size = kwargs.get("batch_size", 1)
 
         img_dim = 64
